unshare utility function
