+++ seqr/views/utils/anvil_metadata_utils.py
@@ -10,8 +10,9 @@ from typing import Callable, Iterable
 from matchmaker.models import MatchmakerSubmission
 from reference_data.models import HumanPhenotypeOntology, Omim, GENOME_VERSION_LOOKUP
 from seqr.models import Project, Family, Individual, Sample, SavedVariant, VariantTagType
-from seqr.views.utils.airtable_utils import get_airtable_samples
+from seqr.views.utils.airtable_utils import AirtableSession
 from seqr.utils.gene_utils import get_genes
+from seqr.utils.middleware import ErrorsWarningsException
 from seqr.utils.search.utils import get_search_samples
 from seqr.utils.xpos_utils import get_chrom_pos
 from seqr.views.utils.variant_utils import DISCOVERY_CATEGORY
@@ -543,11 +544,52 @@ SINGLE_SAMPLE_FIELDS = ['Collaborator', 'dbgap_study_id', 'dbgap_subject_id', 'd
 LIST_SAMPLE_FIELDS = ['SequencingProduct', 'dbgap_submission']
 
 
-def _get_sample_airtable_metadata(sample_ids, user, fields):
-    single_fields, list_fields = fields or [SINGLE_SAMPLE_FIELDS, LIST_SAMPLE_FIELDS]
-    sample_records, _ = get_airtable_samples(
-        sample_ids, user, fields=single_fields, list_fields=list_fields,
+def _get_airtable_samples_for_id_field(sample_ids, id_field, fields, session):
+    raw_records = session.fetch_records(
+        'Samples', fields=[id_field] + fields,
+        or_filters={f'{{{id_field}}}': sample_ids},
     )
+
+    records_by_id = defaultdict(list)
+    for record in raw_records.values():
+        records_by_id[record[id_field]].append(record)
+    return records_by_id
+
+
+def _get_sample_airtable_metadata(sample_ids, user, airtable_fields):
+    fields, list_fields = airtable_fields or [SINGLE_SAMPLE_FIELDS, LIST_SAMPLE_FIELDS]
+    all_fields = fields + list_fields
+
+    session = AirtableSession(user)
+    records_by_id = _get_airtable_samples_for_id_field(sample_ids, 'CollaboratorSampleID', all_fields, session)
+    missing = set(sample_ids) - set(records_by_id.keys())
+    if missing:
+        records_by_id.update(_get_airtable_samples_for_id_field(missing, 'SeqrCollaboratorSampleID', all_fields, session))
+
+    sample_records = {}
+    for record_id, records in records_by_id.items():
+        parsed_record = {}
+        for field in fields:
+            record_field = {
+                record[field][0] if field == 'Collaborator' else record[field] for record in records if field in record
+            }
+            if len(record_field) > 1:
+                error = 'Found multiple airtable records for sample {} with mismatched values in field {}'.format(
+                    record_id, field)
+                raise ErrorsWarningsException([error])
+            if record_field:
+                parsed_record[field] = record_field.pop()
+        for field in list_fields:
+            parsed_record[field] = set()
+            for record in records:
+                if field in record:
+                    values = record[field]
+                    if isinstance(values, str):
+                        values = [values]
+                    parsed_record[field].update(values)
+
+        sample_records[record_id] = parsed_record
+
     return sample_records
